Support custom conditions in Regular
This change allows us to build fixed schedulers which integrate more nicely with telemetry and logging.

Previously, fixed scheduling had to be implemented by regularly starting the job process, and then deciding if we'll actually do something within that process.

This could lead to excessive starting of processes, but more importantly it meant that telemetry and logging would generate a bunch of needless noise.

The new option allows the clients to make the decision directly in the scheduler process, which means that job-related settings, such as `on_overlap`, and `timeout`, as well as telemetry and logging will work as expected.